requirements_mwaa.txt
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.5.1/constraints-3.10.txt"

apache-airflow-providers-amazon==7.1.0
apache-airflow-providers-snowflake==4.0.2
snowflake-connector-python==2.9.0
apache-airflow-providers-github==2.2.0
apache-airflow-providers-google==8.8.0
apache-airflow-providers-cncf-kubernetes==5.1.1
apache-airflow-providers-dbt-cloud==2.3.1
dag-factory==0.18.1
dbt-core>=1.3.0
dbt-postgres>=1.3.0
dbt-redshift>=1.3.0
dbt-extractor==0.4.1